+++ src/frontend/components/organisms/NavBar.js
@@ -1,6 +1,5 @@
 import React, { useCallback } from 'react';
 import { useMappedState } from 'redux-react-hook';
-import { unstable_useMediaQuery as useMediaQuery } from '@material-ui/core/useMediaQuery';
 
 import NavDrawer from '../molecules/NavDrawer';
 import MapToolbar from '../molecules/MapToolbar';
@@ -9,7 +8,6 @@ import NavToolbar from './NavToolbar';
 import AppBar from '@material-ui/core/AppBar';
 
 const NavBar = () => {
-  const large = useMediaQuery('(min-width: 600px)');
   const isMapDetail = useMappedState(
     useCallback(state => state.shared.isMapDetail, [])
   );
