+++ src/frontend/components/organisms/SpotDialog.js
@@ -18,12 +18,6 @@ const styles = {
   appbar: {
     position: 'relative'
   },
-  toolbar: {
-    paddingLeft: 8
-  },
-  flex: {
-    flex: 1
-  },
   dialogContent: {
     padding: 0
   },
@@ -65,11 +59,11 @@ const SpotDialog = () => {
     >
       {!large && (
         <AppBar style={styles.appbar} color="primary">
-          <Toolbar style={styles.toolbar}>
+          <Toolbar>
             <IconButton color="inherit" onClick={handleRequestDialogClose}>
               <ArrowBackIcon />
             </IconButton>
-            <Typography variant="h6" color="inherit" style={styles.flex}>
+            <Typography variant="h6" color="inherit">
               {I18n.t('spot')}
             </Typography>
           </Toolbar>
